edge/analysis_core/requirements.txt
joblib==1.3.1
keras==2.13.1
numpy==1.24.3
paho-mqtt==1.6.1
pandas==2.0.3
pymongo==4.4.0
scikit-learn==1.1.3
scipy==1.10.1
tensorflow==2.13.0
dotenv
